fix: Support different WhatsApp export date formats (#26)
Helpfully WhatsApp for iOS seems to use the system locale for dates. That means our date importer has to be pretty permissive. This change goes same way to addressing this issue, but we'll probably want to allow users to explicitly specify the date format used by their export, and the time zone of the export.